Fix azure unit tests and move to pytest
